fix sequence parallel for dummy forward

+++ csrc/core/memcpy.cu
@@ -66,7 +66,7 @@ __global__ void memcpy_nvshmem_cp(
   const int64_t *seq_tokens,
   const int64_t token_bytes,
   const int32_t total_num_tokens,
-  const int64_t max_num_cp,
+  const int32_t max_num_cp,
   const int64_t num_seq
 ) {
   // each kv replica is a logical sequence.
